// ui/src/services/cradle/models/ReportRenderNotification.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * CRADLE
 * Threat Intelligence Knowledge Management
 *
 * The version of the OpenAPI document: 2.10.2-beta.9bd46310 (2.10.2-beta.9bd46310)
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ReportRenderNotification
 */
export interface ReportRenderNotification {
    /**
     * 
     * @type {string}
     * @memberof ReportRenderNotification
     */
    readonly id?: string;
    /**
     * 
     * @type {string}
     * @memberof ReportRenderNotification
     */
    message: string;
    /**
     * 
     * @type {boolean}
     * @memberof ReportRenderNotification
     */
    isMarkedUnread?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ReportRenderNotification
     */
    readonly timestamp?: Date;
    /**
     * 
     * @type {string}
     * @memberof ReportRenderNotification
     */
    publishedReportId: string;
    /**
     * 
     * @type {string}
     * @memberof ReportRenderNotification
     */
    readonly notificationType?: string;
}

/**
 * Check if a given object implements the ReportRenderNotification interface.
 */
export function instanceOfReportRenderNotification(value: object): value is ReportRenderNotification {
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('publishedReportId' in value) || value['publishedReportId'] === undefined) return false;
    return true;
}

export function ReportRenderNotificationFromJSON(json: any): ReportRenderNotification {
    return ReportRenderNotificationFromJSONTyped(json, false);
}

export function ReportRenderNotificationFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReportRenderNotification {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'message': json['message'],
        'isMarkedUnread': json['is_marked_unread'] == null ? undefined : json['is_marked_unread'],
        'timestamp': json['timestamp'] == null ? undefined : (new Date(json['timestamp'])),
        'publishedReportId': json['published_report_id'],
        'notificationType': json['notification_type'] == null ? undefined : json['notification_type'],
    };
}

export function ReportRenderNotificationToJSON(json: any): ReportRenderNotification {
    return ReportRenderNotificationToJSONTyped(json, false);
}

export function ReportRenderNotificationToJSONTyped(value?: Omit<ReportRenderNotification, 'id'|'timestamp'|'notification_type'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
        'is_marked_unread': value['isMarkedUnread'],
        'published_report_id': value['publishedReportId'],
    };
}